# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for ATS-Tex Lab, a custom CV builder and ATS simulator.",
        "title": "ATS-Tex Lab API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/subscription-plans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all subscription plans (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Subscriptions"
                ],
                "summary": "List Subscription Plans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_semmidev_atstex-lab_internal_domain.SubscriptionPlan"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new subscription plan (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Subscriptions"
                ],
                "summary": "Create Subscription Plan",
                "parameters": [
                    {
                        "description": "Plan data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_semmidev_atstex-lab_internal_domain.SubscriptionPlan"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_semmidev_atstex-lab_internal_domain.SubscriptionPlan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
        "/api/admin/subscription-plans/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing subscription plan (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Subscriptions"
                ],
                "summary": "Update Subscription Plan",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Plan data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_semmidev_atstex-lab_internal_domain.SubscriptionPlan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_semmidev_atstex-lab_internal_domain.SubscriptionPlan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete a subscription plan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Subscriptions"
                ],
                "summary": "Delete Subscription Plan",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
        "/api/admin/subscription-plans/{id}/toggle": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Activate or deactivate a subscription plan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Subscriptions"
                ],
                "summary": "Toggle Subscription Plan Active State",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Toggle status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.AdminToggleSubscriptionPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
...
        "/api/admin/users/{id}/subscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grant a specific subscription plan to a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Subscriptions"
                ],
                "summary": "Assign Subscription to User",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subscription Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.AdminAssignSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
        "/api/admin/users/{id}/subscription": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a user's active subscription",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Subscriptions"
                ],
                "summary": "Get User's Subscription",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_semmidev_atstex-lab_internal_domain.SubscriptionPlan": {
            "type": "object",
            "properties": {
                "activeUsersCount": {
                    "description": "Joined/Computed fields",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "durationMonths": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "maxAtsSimulations": {
                    "type": "integer"
                },
                "maxCoverLetters": {
                    "type": "integer"
                },
                "maxCvProfiles": {
                    "type": "integer"
                },
                "maxCvReviews": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "priceIdr": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_handler.AdminAssignSubscriptionRequest": {
            "type": "object",
            "properties": {
                "months": {
                    "description": "Can be passed as string by HTML forms",
                    "type": "string"
                },
                "planId": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.AdminToggleSubscriptionPlanRequest": {
            "type": "object",
            "properties": {
                "isActive": {
                    "type": "boolean"
                }
            }
        },
...
        "problem.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence of the problem.",
                    "type": "string"
                },
                "instance": {
                    "description": "Instance is a URI reference that identifies the specific occurrence of the problem.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code for this occurrence of the problem.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI reference that identifies the problem type.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}